rename shared to ui
