+++ apps/design/pages/_app.tsx
@@ -1,6 +1,6 @@
 import type { AppProps } from 'next/app'
 import Head from 'next/head'
-import 'shared/globals.css'
+import 'ui/globals.css'
 
 export default function App({ Component, pageProps }: AppProps) {
 	return (
