+++ apps/games/package.json
@@ -17,7 +17,7 @@
 		"next": "13.3.0",
 		"react": "18.2.0",
 		"react-dom": "18.2.0",
-		"shared": "workspace:*"
+		"ui": "workspace:*"
 	},
 	"devDependencies": {
 		"config": "workspace:*",
