+++ apps/games/pages/spider.tsx
@@ -6,8 +6,8 @@ import { ReloadIcon, ResetIcon } from '@radix-ui/react-icons'
 import { isDefined, isEmpty } from 'lib/utils/lang'
 import { classNames } from 'lib/utils/string'
 import { chunk, first, last } from 'lib/utils/array'
-import Header from 'shared/Header'
-import HeaderAction from 'shared/HeaderAction'
+import Header from 'ui/Header'
+import HeaderAction from 'ui/HeaderAction'
 import type { Card, Deck, Piles, Rank, Suit } from '@/types/card-games'
 import { createDeck } from '@/lib/deck'
 import {
