+++ apps/next/app/calculate/page.tsx
@@ -3,7 +3,7 @@
 import { FormEvent, useState } from 'react'
 import Content from '@/components/Content'
 import Section from '@/components/Section'
-import Header from 'shared/Header'
+import Header from 'ui/Header'
 
 const Calculate = () => {
 	const [sum, setSum] = useState(0)
