+++ apps/next/app/confirm-button/page.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react'
 import Content from '@/components/Content'
 import ConfirmButton from '@/components/ConfirmButton'
 import Section from '@/components/Section'
-import Header from 'shared/Header'
+import Header from 'ui/Header'
 
 const ConfirmButtonPage = () => {
 	const [confirmed, setConfirmed] = useState(false)
