+++ apps/next/app/encryption/page.tsx
@@ -3,7 +3,7 @@
 import { useRef, useState } from 'react'
 import { Label } from '@radix-ui/react-label'
 import { AES, enc } from 'crypto-js'
-import Header from 'shared/Header'
+import Header from 'ui/Header'
 import Content from '@/components/Content'
 import Section from '@/components/Section'
