+++ apps/next/app/ms/page.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useState } from 'react'
 import Section from '@/components/Section'
-import Header from 'shared/Header'
+import Header from 'ui/Header'
 import Content from '@/components/Content'
 import MineSweaper from '@/components/MineSweaper'
