+++ apps/next/app/password-generator/page.tsx
@@ -4,11 +4,11 @@ import { NextPage } from 'next'
 import { PropsWithChildren, useState } from 'react'
 import * as Tabs from '@radix-ui/react-tabs'
 import useSWR from 'swr'
-import { Slider } from 'shared/Slider'
+import { Slider } from 'ui/Slider'
 import { randomString } from 'lib/utils/string'
-import { CheckboxWithLabel } from 'shared/CheckboxWithLabel'
+import { CheckboxWithLabel } from 'ui/CheckboxWithLabel'
 import Content from '@/components/Content'
-import Header from 'shared/Header'
+import Header from 'ui/Header'
 import Section from '@/components/Section'
 
 interface Selection {
