+++ apps/next/app/two-way-auth-enter/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import TwoWayAuthEnter from '@/components/TwoWayAuthEnter'
-import Header from 'shared/Header'
+import Header from 'ui/Header'
 import Content from '@/components/Content'
 import Section from '@/components/Section'
