+++ apps/next/package.json
@@ -32,7 +32,7 @@
 		"react": "18.2.0",
 		"react-dom": "18.2.0",
 		"react-icons": "4.7.1",
-		"shared": "workspace:*",
+		"ui": "workspace:*",
 		"swr": "1.3.0"
 	},
 	"devDependencies": {
