+++ apps/password-manager/pages/_app.tsx
@@ -1,7 +1,7 @@
 import type { AppProps } from 'next/app'
 import { UserProvider } from '@supabase/auth-helpers-react'
 import { supabaseClient } from '@supabase/auth-helpers-nextjs'
-import 'shared/globals.css'
+import 'ui/globals.css'
 
 function MyApp({ Component, pageProps }: AppProps) {
 	return (
