+++ packages/ui/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "shared",
+	"name": "ui",
 	"version": "0.0.0",
 	"private": true,
 	"scripts": {
